Create structure BNF
Added BNF grammar for structures (alias and record). Additionally, also
updated grammar for attribute list and symbols.

The purpose of !(<<eof>>) and !(<<eof>> | 'ENDMODULE') is to stop
expecting '<TDN>', '<DDN>', '<RDN>', '<symbol>', or '(' when actually
expecting 'ENDMODULE'.

Symbol is not marked as private, but all structures, fields, and
routines should replace it.