Add functionality to download and edit remote modules
Modules are saved by the Module entity. The file structure is organized in RobotUtil, where 3 methods are added, one to download modules from a remote robot, one to fetch persisted modules, and one to remove persisted modules. Currently, all previously persisted modules are removed on refresh, which causes all unsaved changes to be discarded, as a result, a confirmation should be added.

Tasks and their respective modules are also visible, and navigable, through the tool window.

Furthermore, persisted modules are scanned for changes, and any changes lead to the remote module being updated accordingly.