Update RobotConnectView UI
Update UI for RobotConnectView.

A notification is displayed to the user if a network request fails, with a link to open RobotConnectView, with the previous host and port added.

Fixed an issue with RobotUtil#getRobotState where if a request failed the exception was not forwarded and the calling thread was not released.