Add I/O service
Add service used to connect to I/O system of robot. The service is used to collect I/O symbols for symbol resolution. Almost every endpoint is mapped, except for some configuration endpoints, search endpoints, and the ability to set a value to a signal. The issue with these endpoints is the use of optional endpoints.