Build ListProxy content non-lazily
Build elements in a ListProxy immediately, and throw an IOException or
InterruptedException if an error occurs. Currently, the list is built
semi-lazily, as the entire list is built as soon as an element needs to
be retrieved. However, if this operation fails, a ProxyException is
thrown, which is difficult to catch, since it made clear.