Provide fix to change variable type to provided type
Provide a fix to change the variable type to the provided type, if the
left and right expressions in an assignment statement are not
assignable.

Additionally, provide a fix to change the return type of a routine, if
the type of the expression being returned and the routine type are not
assignable.